{"schema":"decodex/commit/1","summary":"Modularize repo gate git operations","authority":"manual"}
